Solve some unix system build issues
Solve some issues when building this plugin on unix system.

1)
The repository does not correctly set the executable flag on the gradlew
file, preventing execution of the file via e.g. the './' syntax.

2)
Correct the casing of the folder name for the 1.20.6 adapter, which
incorrectly used a lowercase 'v', while the settings.gradle.kts referred
to it via an uppercase 'V'. This faulty configuration is not problematic
on case-insensitive file systems like NTFS, however completely breaks
building on case-sensitive file systems like ext4.